refactor(api): split permissions.js controller into per-responsibility modules (#599)
permissions.js had grown to 847 lines (audit finding C1). Split it by
responsibility into focused routers under routes/permissions/:

  grid.js          — /api/permissions + /api/user-columns (the matrix grid feed + its filter columns)
  accessPackages.js — /api/access-package-groups (+ /api/access-package-resources alias)
  syncLog.js       — /api/sync-log
  nestedGroups.js  — /api/groups-with-nested + /api/group/:id/nested-groups
  shared.js        — the useSql flag + the lazily-imported db module

permissions.js is now a barrel that composes the four routers, so app.js
and every public path are unchanged. Pure code move.

Grandfathered complexity entries are relocated (no increase) to the new
paths: cyclomatic [46] -> grid; cognitive [104,28,18,17] -> grid [104,28]
+ accessPackages [18] + syncLog [17].

Added permissions.mock.test.js covering the USE_SQL=false mock-data and
no-database branches that the split moved — diff coverage of the changed
lines is 91% (unit-only), clearing the 80% gate. Full API suite green
(1740 tests); coverage above every floor; both complexity gates and the
OpenAPI-drift guard pass.
